:recycle: Split then steps into modules
